clean up varbin array module
