+++ vortex-array/src/arrays/varbin/vtable/operations.rs
