+++ vortex-array/src/arrays/varbin/vtable/visitor.rs
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use crate::arrays::{VarBinArray, VarBinVTable};
+use crate::vtable::{ValidityHelper, VisitorVTable};
+use crate::{ArrayBufferVisitor, ArrayChildVisitor};
+
+impl VisitorVTable<VarBinVTable> for VarBinVTable {
+    fn visit_buffers(array: &VarBinArray, visitor: &mut dyn ArrayBufferVisitor) {
+        visitor.visit_buffer(array.bytes()); // TODO(ngates): sliced bytes?
+    }
+
+    fn visit_children(array: &VarBinArray, visitor: &mut dyn ArrayChildVisitor) {
+        visitor.visit_child("offsets", array.offsets());
+        visitor.visit_validity(array.validity(), array.len());
+    }
+}
