+++ vortex-array/src/arrays/varbinview/compute/min_max.rs
@@ -3,13 +3,13 @@
 
 use vortex_error::VortexResult;
 
-use crate::arrays::{VarBinViewArray, VarBinViewVTable, compute_min_max};
+use crate::arrays::{VarBinViewArray, VarBinViewVTable, varbin_compute_min_max};
 use crate::compute::{MinMaxKernel, MinMaxKernelAdapter, MinMaxResult};
 use crate::register_kernel;
 
 impl MinMaxKernel for VarBinViewVTable {
     fn min_max(&self, array: &VarBinViewArray) -> VortexResult<Option<MinMaxResult>> {
-        compute_min_max(array, array.dtype())
+        varbin_compute_min_max(array, array.dtype())
     }
 }
